refactor(graphql): Remove AbstractCodepoint class
The Codepoint type is now an abstract class and directly implements the
common behavior and attributes.
