+++ src/main/java/org/unicodetool/graphql/schema/Character.java
@@ -3,7 +3,7 @@ package org.unicodetool.graphql.schema;
 /**
  * Codepoint that encodes an assigned character
  */
-public class Character extends AbstractCodepoint {
+public class Character extends Codepoint {
     public Character(CodepointValue value, String name, Properties props) {
         super(value, name, props);
     }
