+++ src/main/java/org/unicodetool/graphql/schema/Codepoint.java
@@ -1,9 +1,31 @@
 package org.unicodetool.graphql.schema;
 
-public interface Codepoint {
-    CodepointValue getValue();
-    Integer getDecimalValue();
-    String getName();
+import lombok.AllArgsConstructor;
 
-    Properties getProperties();
+/**
+ * Common behaviour between all codepoint types. Corresponds to the Codepoint Interface in the schema.
+ */
+@AllArgsConstructor
+public abstract class Codepoint {
+    private CodepointValue value;
+    private String name;
+
+    private Properties properties;
+
+    public CodepointValue getValue() {
+        return value;
+    }
+
+    public Integer getDecimalValue() {
+        return value.inDecimalFormat();
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public Properties getProperties() {
+        return properties;
+    }
+
 }
