+++ src/main/java/org/unicodetool/graphql/schema/NonCharacter.java
@@ -3,7 +3,7 @@ package org.unicodetool.graphql.schema;
 /**
  * Codepoint that encodes a noncharacter.
  */
-public class NonCharacter extends AbstractCodepoint {
+public class NonCharacter extends Codepoint {
     public NonCharacter(CodepointValue value, String name, Properties properties) {
         super(value, name, properties);
     }
