+++ src/main/java/org/unicodetool/graphql/schema/Reserved.java
@@ -3,7 +3,7 @@ package org.unicodetool.graphql.schema;
 /**
  * Codepoint that is reserved for future use.
  */
-public class Reserved extends AbstractCodepoint {
+public class Reserved extends Codepoint {
     public Reserved(CodepointValue value, String name, Properties properties) {
         super(value, name, properties);
     }
