+++ src/main/java/org/unicodetool/graphql/schema/Surrogate.java
@@ -3,7 +3,7 @@ package org.unicodetool.graphql.schema;
 /**
  * Represent a surrogate codepoint used in surrogate pairs for UTF-16
  */
-public class Surrogate extends AbstractCodepoint {
+public class Surrogate extends Codepoint {
     public Surrogate(CodepointValue value, String name, Properties properties) {
         super(value, name, properties);
     }
